# docs/swagger.yaml
basePath: /
definitions:
  account.AccountStatusUpdate:
    properties:
      status:
        example: BLOCKED
        type: string
    required:
    - status
    type: object
  account.Deposit:
    properties:
      account_no:
        example: ACC87112088
        type: string
      amount:
        example: 500
        type: number
    required:
    - account_no
    - amount
    type: object
  account.TransferInput:
    properties:
      amount:
        example: 500
        type: number
      receiver_acc:
        example: ACC34211234
        type: string
      sender_acc:
        example: ACC34218765
        type: string
    required:
    - amount
    - receiver_acc
    - sender_acc
    type: object
  account.Withdraw:
    properties:
      account_no:
        example: ACC34218765
        type: string
      amount:
        example: 500
        type: number
    required:
    - account_no
    - amount
    type: object
  response.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  user.LoginInput:
    properties:
      email:
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  user.RegistrationInput:
    properties:
      email:
        example: mohsin@example.com
        type: string
      name:
        example: Mohsin Shanto
        type: string
      password:
        example: password123
        maxLength: 32
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: A Banking REST API built with Gin and GORM.
  title: Banking API
  version: "1.0"
paths:
  /account/:
    post:
      description: Creates a bank account for the authenticated user.
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Account already exists
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create a new bank account
      tags:
      - Account
  /account/deposit:
    post:
      consumes:
      - application/json
      description: Deposit money into an active bank account.
      parameters:
      - description: Deposit Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.Deposit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Deposit money into an account
      tags:
      - Account
  /account/details/{accountNo}:
    get:
      description: Retrieve details of a bank account by account number.
      parameters:
      - description: Account Number
        in: path
        name: accountNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account details retrieved successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get account details
      tags:
      - Account
  /account/statistics/{accountNo}:
    get:
      description: Retrieve transaction statistics for a specific bank account.
      parameters:
      - description: Account Number
        in: path
        name: accountNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction statistics retrieved successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get transaction statistics
      tags:
      - Account
  /account/status/{accountNo}:
    put:
      consumes:
      - application/json
      description: Update the status of an existing bank account.
      parameters:
      - description: Account Number
        in: path
        name: accountNo
        required: true
        type: string
      - description: Account Status Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.AccountStatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Account status updated successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Invalid request, invalid status, or status already set
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update account status
      tags:
      - Account
  /account/summary/{accountNo}:
    get:
      description: Retrieve a summary of a bank account, including balance and transaction
        totals.
      parameters:
      - description: Account Number
        in: path
        name: accountNo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account summary retrieved successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get account summary
      tags:
      - Account
  /account/transactions/{accountNo}:
    get:
      description: Retrieve transactions for a specific account with pagination, filtering,
        date range, and sorting.
      parameters:
      - description: Account Number
        in: path
        name: accountNo
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 5
        description: Number of records per page
        in: query
        name: limit
        type: integer
      - description: Transaction type (DEPOSIT, WITHDRAW, TRANSFER)
        in: query
        name: type
        type: string
      - description: Start date (YYYY-MM-DD), e.g. 2026-08-01
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD), e.g. 2026-08-31
        in: query
        name: to
        type: string
      - default: newest
        description: Sort order (newest, oldest)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transactions retrieved successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get account transactions
      tags:
      - Account
  /account/transfer:
    post:
      consumes:
      - application/json
      description: Transfer money from one active account to another active account.
      parameters:
      - description: Money Transfer Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.TransferInput'
      produces:
      - application/json
      responses:
        "200":
          description: Money transferred successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Invalid request, inactive account, insufficient balance, or
            same account transfer
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Transfer money between two accounts
      tags:
      - Account
  /account/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraw money from an active bank account
      parameters:
      - description: Withdraw Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.Withdraw'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Withdraw money from an account
      tags:
      - Account
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, then generate a JWT
        token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Token generated successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Login user
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegistrationInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      summary: Register a new user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"